Add Thread safety for HMAC_DRBG Refactorization of code for removing Utils::zeroize dependancies

+++ include/crypto/Utils.hpp
@@ -21,8 +21,8 @@ namespace Crypto
 			static int from_hex(const std::string, uint8_t*, std::size_t&);
 			static int to_hex(const uint8_t*, std::size_t, std::string&, bool=true);
 
-			static const uint8_t CRYPTO_UTILS_SUCCESS          = 0x00;
-			static const uint8_t CRYPTO_UTILS_INCORRECT_LENGTH = 0x01;
+			static const int CRYPTO_UTILS_SUCCESS          = 0x00;
+			static const int CRYPTO_UTILS_INCORRECT_LENGTH = 0x01;
 
 			class Exception : public std::runtime_error
 			{
